refactor: 메인 페이지 컴포넌트들 분리
